SNAKE-44 fix store
