--- src/ssrStore/reducers/endGame/index.ts
@@ -1,4 +1,0 @@
-export * from './actions';
-export * from './reducer';
-export * from './selectors';
-export * from './types';
