--- src/ssrStore/reducers/index.ts
@@ -1,4 +1,0 @@
-export * from './rootReducer';
-export * from './user';
-export * from './modal';
-export * from './endGame';
