--- src/ssrStore/reducers/modal/index.ts
@@ -1,4 +1,0 @@
-export * from './reducer';
-export * from './selectors';
-export * from './actions';
-export * from './types';
