--- src/ssrStore/types.ts
@@ -1,7 +1,0 @@
-import { configureStore } from './configureStore';
-
-export type SsrState = ReturnType<ReturnType<typeof configureStore>['store']['getState']>;
-
-export interface BaseAction<T> {
-  type: T;
-}
