+++ src/store/getInitialState.ts
