+++ src/store/types.ts
@@ -1,6 +1,6 @@
-import { store } from './configureStore';
+import { configureStore } from './configureStore';
 
-export type RootState = ReturnType<typeof store.getState>;
+export type RootState = ReturnType<ReturnType<typeof configureStore>['store']['getState']>;
 
 export interface BaseAction<T> {
   type: T;
